refactoring stop instance, giving feedback to stdio

+++ src/utils/util.c
@@ -81,7 +81,7 @@ static void close_file_descriptor(FILE *fd)
         }
 }
 
-void check_if_running()
+int is_running()
 {
         char *cmd_pgrep = "pgrep";
         char *cmd_cnt = "| wc -l";
@@ -90,35 +90,41 @@ void check_if_running()
         sprintf(cmd, "%s %s %s", cmd_pgrep, _NAME_, cmd_cnt);
 
         FILE *fd = command_file_descriptor_exec(cmd);
-        
+
         char cmdo[MAX_PID_LEN];
         fgets(cmdo, MAX_PID_LEN, fd);
         int nprcss = atoi(cmdo);
 
         close_file_descriptor(fd);    
 
-        if (nprcss > 1) {
-                fprintf(stderr, "there is another instance of canaryusb running\n");
-                exit(EXIT_FAILURE);
-        }
+        if (nprcss > 1)
+                return 1;
+        else
+                return 0;
 }
 
 void kill_canaryusb_inst()
 {
-        char *cmd_pgrep = "kill $(pgrep";
-        char *cmd_cnt = "| head -1)";
-        char *cmd = (char *) malloc(strlen(cmd_pgrep) + strlen(_NAME_) + strlen(cmd_cnt) + 2);
-        check_memory_allocation(cmd);
-        sprintf(cmd, "%s %s %s", cmd_pgrep, _NAME_, cmd_cnt);
+       if (is_running()) {
+               char *cmd_pgrep = "kill $(pgrep";
+               char *cmd_cnt = "| head -1)";
+               char *cmd = (char *) malloc(strlen(cmd_pgrep) + strlen(_NAME_) + strlen(cmd_cnt) + 2);
+               check_memory_allocation(cmd);
+               sprintf(cmd, "%s %s %s", cmd_pgrep, _NAME_, cmd_cnt);
 
-        FILE *fd = command_file_descriptor_exec(cmd);
-        
-        char cmdo[MAX_PID_LEN];
-        fgets(cmdo, MAX_PID_LEN, fd);
-        close_file_descriptor(fd);    
+               FILE *fd = command_file_descriptor_exec(cmd);
+               
+               char cmdo[MAX_PID_LEN];
+               fgets(cmdo, MAX_PID_LEN, fd);
+               close_file_descriptor(fd);    
 
-        syslog(LOG_NOTICE, "%s daemon stopped", _NAME_);
-        exit(EXIT_SUCCESS);
+               syslog(LOG_NOTICE, "%s daemon stopped", _NAME_);
+               printf("%s daemon stopped\n", _NAME_);
+       } else {
+               printf("no %s daemon running, nothing to stop\n", _NAME_);
+       }
+       
+       exit(EXIT_SUCCESS);
 }
 
 void check_argument_length(char *arg, int type)
